+++ cmd/awoocc/lexer_token/lexer_token.go
@@ -4,15 +4,15 @@ import "github.com/LamkasDev/awoo-emu/cmd/awoocc/token"
 
 type AwooLexerToken struct {
 	Type  uint16
-	Start uint16
+	Start uint32
 	Data  interface{}
 }
 
 type FetchToken func() (AwooLexerToken, error)
 
-func CreateToken(start uint16, t *token.AwooToken) AwooLexerToken {
+func CreateToken(start uint32, t *token.AwooToken) AwooLexerToken {
 	return AwooLexerToken{
 		Type:  t.Id,
-		Start: start - uint16(t.Length) + 1,
+		Start: start - uint32(t.Length) + 1,
 	}
 }
