+++ cmd/awoocc/lexer_token/primitive.go
@@ -35,7 +35,7 @@ func SetTokenPrimitiveExtra(t *AwooLexerToken, value interface{}) {
 	t.Data.(*AwooLexerTokenDataPrimitive).Extra = value
 }
 
-func CreateTokenPrimitive(start uint16, primitiveType types.AwooTypePrimitiveId, value interface{}, extra interface{}) AwooLexerToken {
+func CreateTokenPrimitive(start uint32, primitiveType types.AwooTypePrimitiveId, value interface{}, extra interface{}) AwooLexerToken {
 	return AwooLexerToken{
 		Type:  token.TokenTypePrimitive,
 		Start: start,
