+++ cmd/awoocc/lexer_token/type.go
@@ -17,7 +17,7 @@ func SetTokenTypeId(t *AwooLexerToken, id types.AwooTypeId) {
 	t.Data.(*AwooLexerTokenDataType).Id = id
 }
 
-func CreateTokenType(start uint16, value types.AwooTypeId) AwooLexerToken {
+func CreateTokenType(start uint32, value types.AwooTypeId) AwooLexerToken {
 	return AwooLexerToken{
 		Type:  token.TokenTypeType,
 		Start: start,
